requirements.txt
pandas
numpy
matplotlib
plotly
streamlit
sqlalchemy
fredapi
statsmodels
prophet
python-dotenv
